Press SELECT on the durationspinbox to start the timer
removes one more keystroke to start a timer. It is now possible to press
select while the duration spinbox is selected to start the timer.

Thus it is now possible to open the timer editor, move with one
keystroke ("UP") to the timer duration spinbox, change to the desired
duration and immediatly press "SELECT" to start the timer and leave the
plugin dialogs.